Support UDP packet fragmentation
Under "native" mode the max size of UDP packets is no longer capped by
QUIC's max datagram size and is now only determined by the value in
config. This is done by adding a LongPacket command along side Packet
which contains fragmentation info (the protocol is hence upgraded to v5
but is still backwards compatible).

Testing shows that the first couple of packets may be dropped silently
(which is acceptable for UDP) even on stable connections, but that's
acceptable for UDP.